# docs/openapi/chronos-api.yaml
openapi: 3.1.0
info:
  title: CHRONOS INFINITY API
  version: 2.0.0
  description: |
    API empresarial de gestión financiera para el sistema CHRONOS INFINITY 2026.

    ## Características
    - Gestión de ventas con distribución automática GYA
    - Control de bancos/bóvedas financieras
    - Tracking de clientes y distribuidores
    - Órdenes de compra con trazabilidad
    - Reportes y analytics en tiempo real

    ## Autenticación
    La API usa JWT Bearer tokens para autenticación.

    ## Rate Limiting
    - 100 requests/minuto por IP
    - 1000 requests/hora por usuario autenticado
  contact:
    name: CHRONOS Support
    email: support@chronos.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.chronos.app/v2
    description: Production
  - url: https://staging-api.chronos.app/v2
    description: Staging
  - url: http://localhost:3000/api
    description: Development
...
paths:
...
  /clientes:
    get:
      summary: Listar clientes
      tags: [Clientes]
      operationId: getClientes
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: search
          in: query
          schema:
            type: string
          description: Búsqueda por nombre o contacto
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cliente'

    post:
      summary: Crear cliente
      tags: [Clientes]
      operationId: createCliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearClienteInput'
      responses:
        '201':
          description: Cliente creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    Cliente:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nombre:
          type: string
        contacto:
          type: string
        email:
          type: string
          format: email
        telefono:
          type: string
        direccion:
          type: string
        createdAt:
          type: string
          format: date-time
    CrearClienteInput:
      type: object
      required: [nombre]
      properties:
        nombre:
          type: string
          minLength: 2
          maxLength: 100
        contacto:
          type: string
        email:
          type: string
          format: email
        telefono:
          type: string
        direccion:
          type: string
...
security:
  - BearerAuth: []